requirements.txt
python-telegram-bot[job-queue]
python-dotenv
libsql-client
feedparser
requests
apscheduler
